fix cubemap texture
